Move test utilities where they are needed
Sorry for moving files while changing them at
the same time. This was necessary for keeping
the tests pass.
