+++ common/connection_test.go
@@ -30,14 +30,14 @@ import (
 	"github.com/chromedp/cdproto/cdp"
 	"github.com/chromedp/cdproto/target"
 	"github.com/gorilla/websocket"
-	"github.com/grafana/xk6-browser/testutils"
+	"github.com/grafana/xk6-browser/common/test"
 	"github.com/mailru/easyjson"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestConnection(t *testing.T) {
-	server := testutils.NewWSTestServerWithEcho(t)
+	server := test.NewWSTestServerWithEcho(t)
 	defer server.Cleanup()
 
 	t.Run("connect", func(t *testing.T) {
@@ -52,7 +52,7 @@ func TestConnection(t *testing.T) {
 }
 
 func TestConnectionClosureAbnormal(t *testing.T) {
-	server := testutils.NewWSTestServerWithClosureAbnormal(t)
+	server := test.NewWSTestServerWithClosureAbnormal(t)
 	defer server.Cleanup()
 
 	t.Run("closure abnormal", func(t *testing.T) {
@@ -70,7 +70,7 @@ func TestConnectionClosureAbnormal(t *testing.T) {
 }
 
 func TestConnectionSendRecv(t *testing.T) {
-	server := testutils.NewWSTestServerWithCDPHandler(t, testutils.CDPDefaultHandler, nil)
+	server := test.NewWSTestServerWithCDPHandler(t, test.CDPDefaultHandler, nil)
 	defer server.Cleanup()
 
 	t.Run("send command with empty reply", func(t *testing.T) {
@@ -134,7 +134,7 @@ func TestConnectionCreateSession(t *testing.T) {
 		}
 	}
 
-	server := testutils.NewWSTestServerWithCDPHandler(t, handler, &cmdsReceived)
+	server := test.NewWSTestServerWithCDPHandler(t, handler, &cmdsReceived)
 	defer server.Cleanup()
 
 	t.Run("create session for target", func(t *testing.T) {
