+++ common/test/cdp.go
@@ -19,7 +19,7 @@
  */
 
 // Package testutils is indended only for use in tests, do not import in production code!
-package testutils
+package test
 
 import (
 	"fmt"
