+++ tests/es6.go
@@ -19,7 +19,7 @@
  */
 
 // Package testutils is indended only for use in tests, do not import in production code!
-package testutils
+package tests
 
 import (
 	"testing"
